+++ sdk/js/src/service/devices/index.js
@@ -19,18 +19,18 @@ import traverse from 'traverse'
 
 import { notify, EVENTS } from '../../api/stream/shared'
 import Marshaler from '../../util/marshaler'
-import combineStreams from '../../util/combine-streams'
+import subscribeToStream from '../../api/stream/subscribeToStream'
 import deviceEntityMap from '../../../generated/device-entity-map.json'
 import DownlinkQueue from '../downlink-queue'
 import { STACK_COMPONENTS_MAP } from '../../util/constants'
 import DeviceClaim from '../claim'
+import combineStreams from '../../util/combine-streams'
 
 import Repository from './repository'
 import { splitSetPaths, splitGetPaths, makeRequests } from './split'
 import mergeDevice from './merge'
 
-const { is: IS, ns: NS, as: AS, js: JS, dtc: DTC } = STACK_COMPONENTS_MAP
-
+const { is: IS, ns: NS, as: AS, js: JS, gs: GS } = STACK_COMPONENTS_MAP
 /**
  * Devices Class provides an abstraction on all devices and manages data
  * handling from different sources. It exposes an API to easily work with
@@ -645,7 +645,7 @@ class Devices {
 
           const result = await this.create(applicationId, end_device, paths)
 
-          notify(listeners[EVENTS.CHUNK], result)
+          notify(listeners[EVENTS.MESSAGE], result)
           finishedCount++
         } catch (error) {
           notify(listeners[EVENTS.ERROR], error)
@@ -693,17 +693,13 @@ class Devices {
     // Event streams can come from multiple stack components. It is necessary to
     // check for stack components on different hosts and open distinct stream
     // connections for any distinct host if need be.
-    const distinctComponents = this._stackConfig.getComponentsWithDistinctBaseUrls([
-      IS,
-      JS,
-      NS,
-      AS,
-      DTC,
-    ])
+    const distinctComponents = this._stackConfig.getComponentsWithDistinctBaseUrls([IS, GS])
 
-    const streams = distinctComponents.map(component =>
-      this._api.Events.Stream({ component }, payload),
+    const baseUrls = new Set(
+      distinctComponents.map(component => this._stackConfig.getComponentUrlByName(component)),
     )
+
+    const streams = [...baseUrls].map(baseUrl => subscribeToStream(payload, baseUrl))
 
     // Combine all stream sources to one subscription generator.
     return combineStreams(streams)